OneDrive/Desktop/bid-command/requirements.txt
PyMuPDF>=1.25.0
openpyxl>=3.1.0
reportlab>=4.0
python-docx>=1.0
pytest>=8.0
